8주차 과제 추가 수정

+++ src/main/java/com/springboot/domain/category/entity/CategoryRepository.java
@@ -7,7 +7,7 @@ import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
-public interface CategoryRepository extends JpaRepository<Category, BigInteger> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
     List<Category> findAllByMemberId(Long memberId);
